优化peer及接口的分类文件
